utils: rename kpf.py -> kpf_utils.py
Disambiguate the module name from the kpfpipe package itself; update all
importers and doc references. Pure rename, no code change.
