+++ kpfpipe/utils/io.py
@@ -11,7 +11,7 @@ import pandas as pd
 from astropy.io import fits
 
 from kpfpipe.utils.config import ConfigHandler
-from kpfpipe.utils.kpf import (
+from kpfpipe.utils.kpf_utils import (
     get_datecode,
     get_obs_id,
     get_timestamp,
