+++ scripts/processing/rv_timeseries.py
@@ -43,7 +43,7 @@ from astropy.io import fits
 import kpfpipe
 from kpfpipe.utils.config import ConfigHandler
 from kpfpipe.utils.io import FileHandler, kpf_filepath
-from kpfpipe.utils.kpf import get_datecode, get_obs_id, is_datecode
+from kpfpipe.utils.kpf_utils import get_datecode, get_obs_id, is_datecode
 
 # Masters every science frame depends on, as (cal_type, level). Flat masters are
 # scaffolded but not implemented, so they are not (yet) required -- this matches
